+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Update some dependencies (`classnames`, `emoji-regex`, `mime-types`, `use-debounce`) and remove unused ones (`immutability-helper`, `@blueprintjs/select`, `wolfy87-eventemitter`, `@types/classnames`, `@types/css`, `@types/sass`, `depcheck`, `pngjs`)
 - use core method to validate email addresses, this will make tld email addresses possible
 - upgrade minimum nodejs to version `14`
+- remove dependencies that are now not needed anymore (`fs-extra`)
 
 ## [1.20.3] - 2021-06-30